requirements.txt
streamlit
pandas
bcrypt
faiss-cpu
sentence-transformers
openai
numpy
kagglehub
groq
tqdm
torch
transformers
scikit-learn
openai-whisper
FFmpeg
pydub